+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 | Release | Date       | Comments                                                           |
 |---------|------------|--------------------------------------------------------------------|
-| 1.205 | 2023.10.15 | Fixed wrong path for server's private keys |
+| 1.22.00 | 2023.11.02 | fixed broken cert rm (duplicated path in pathname                  |
+| 1.21.00 | 2023.11.02 | defaultEnv.json does not need to be specified anymore in cm env    |
+| 1.20.06 | 2023.10.31 | Error message in cert verify less confusing                        |
+| 1.20.05 | 2023.10.16 | New version numbering scheme                                       |
+| 1.205   | 2023.10.15 | Fixed wrong path for server's private keys                         |
 | 1.200   | 2023.10.13 | Go version bump to 1.21.3, simplified path handling in environment |
 | 1.100   | 2023.10.04 | `cm cert verify` now displays comments when asked to               |
 | 1.010-1 | 2023.10.03 | Serial number now properly incremented                             |